[ENH] add ExternalCollectionReadContext to MeterEvent enum (#5320)
## Description of changes

We need to detect when tenants perform read operations on collections
that are not theirs. To achieve this, we introduce an
`ExternalCollectionRead` variant on the `MeterEvent` enum. For the get,
query, and count paths, we conditionally create this context if the
requesting tenant does not match the tenant that owns the requested
resource. In the flushing logic, we first attempt to flush the current
`CollectionReadContext`, and if that fails, we attempt to flush the new
`ExternalCollectionReadContext`.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

This will break hosted. The plan is to merge this PR and then shortly
after merge the corresponding hosted PR:

## Observability plan

N/A

## Documentation Changes

N/A