[ENH]  Make a describe-all.txt that describes the cluster state after testing. (#5359)
## Description of changes

I suspect that our test add failures are because of memory pressure
causing a pod to crash.  I suspect this because I've seen failures to
connect to etcd and the kind cluster, which shouldn't happen once tilt
believes itself to be up and alive.

This PR dumps kubectl describe -A all to get stuff like pod restarts
into the test data that we log per test run.

## Test plan

CI

## Migration plan

N/A

## Observability plan

Look for the file.

## Documentation Changes

N/A