[ENH]: Reduce default max elements (#5361)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- The default max elements param controls how many elements the hnsw
index is inited with. For small indexes, the overhead of this large
allocation is non-trivial as it leads to reading bigger files from s3
which in turn impact the query latency. This PR reduces it from 10k to
100.
- The downside of this is that for large indexes, there will be more
allocation/data movement. But this is in the compaction path which
happens in the background so ok to pay the cost.
- Another consideration is that with SPANN as the default index, hnsw
indexes will be smaller since it only stores the centers now as opposed
to all the data points
- New functionality
  - ...

## Test plan
_How are these changes tested?_
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
_Are there any migrations, or any forwards/backwards compatibility
changes needed in order to make sure this change deploys reliably?_
None

## Observability plan
_What is the plan to instrument and monitor this change?_
Monitor the impact on a user in prod

## Documentation Changes
_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_
None