[ENH]  Cut out the go log service. (#5369)
## Description of changes

This PR will make the push_logs call initialize the log.  Existing logs
are unaffected.  If a log doesn't exist, a not_found error will be
returned or the request will be empty, as appropriate.

This deletes the paths that push/pull logs in the go log service.

## Test plan

CI

## Migration plan

This is part of the migration to the rust log service.

We need to plan for how to roll it out such that the rust log service
rolls first.

## Observability plan

Watch staging.

## Documentation Changes

N/A