[ENH][log]  Cache the manifest and etag. (#5368)
## Description of changes

Prior to this commit, the code cached just the manifest.  By caching the
etag we can HEAD the manifest if it hasn't changed and get some latency
improvements for scout logs.

## Test plan

CI

## Migration plan

I changed the cache key, so even if we had a disk cache it would be
backwards-compatible.

## Observability plan

N/A

## Documentation Changes

N/A

---------

Co-authored-by: propel-code-bot[bot] <203372662+propel-code-bot[bot]@users.noreply.github.com>