[BUG] Reduce RFE metric cardinality (#5395)
## Description of changes

Adding the "tenant_id" attribute to the metrics causes an excessive
number of events to be emitted.

This is a continuation of the effort from
https://github.com/chroma-core/chroma/pull/5373. Turns out, to reduce
event volume even further, we want to omit the attribute for
`tenatnt_id` (in addition to removing the attribute for
`collection_id`).

## Test plan

Unit tests passing, otherwise untested. Manually inspected our metrics
dashboards to ensure we aren't currently using these attributes.

## Migration plan

N/A

## Observability plan

Ensure dashboards remain unaffected.

## Documentation Changes

N/A