[BUG]  Proxy wal3's backoff error without log contention. (#5410)
## Description of changes

Prior to this fix, a backoff error generated from pushing too many logs
would generate the scary, "log contention and your data may or may not
be durable."  Tested that it now generates a Backoff properly.  Then,
make Backoff map to a 429, not a 5xx.

## Test plan

Manual testing + CI

## Migration plan

N/A

## Observability plan

N/A

## Documentation Changes

N/A