[ENH]  A control interface for GC. (#5218)
## Description of changes

This introduces an endpoint for garbage collecting collections from the
command line.  The goal is to inject a collection provided on the
command line in the list of collections to clean up.

Once provided such a collection, gc will collect it at most once.

## Test plan

CI

## Migration plan

N/A

## Observability plan

I plan to add the endpoint and then observe the world as I use it.

## Documentation Changes

N/A