[CHORE] remove tilt_config.yaml (#5440)
## Description of changes

As part of the rust log service migration we introduced a new config.
This turned out to be unnecessary and left us in an unclean state.
Change the tilt config back to distributed.yaml and chroma_config.yaml,
making them identical content-wise to the tilt_config.yaml.

## Test plan

- [X] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

CI

## Migration plan

N/A

## Observability plan

N/A

## Documentation Changes

N/A