[BUG] Remove `rc/**-**-**` branch trigger-deploy workflow (#5465)
## Description of changes

Currently, this workflow triggers on pushes to `rc/**-**-**` branches.
However, in the `_deploy.yaml` workflow (used by this one), the default
values are used for oss-ref and hosted-ref (which is latest and main).
Therefore, pushing to a `rc/**-**-**` branch will end up deploying
latest oss and main hosted, which is not the intended behavior.

For now, it's probably sufficient to require that we manually trigger
deploys after pushing to `rc/**-**-**` branches, instead of hooking up
automation that does the deploy automatically for us.

## Test plan

These changes are not tested, but hopefully relatively safe (given that
we have a manual process for triggering deploys). This is just removing
some (incorrectly behaving) automation.

## Migration plan

n/a

## Observability plan

n/a

## Documentation Changes

TODO, update runbooks.