[CHORE]  Box all large futures. (#5467)
## Description of changes

In CI we are starting to see failures as we add metrics to the leaf
futures.  Every future forms a tree going from the entry point to the
leaf and the entry point must have space to allocate within it every
future to be allocated on the stack.  If there are large futures, they
will bloat the outer future.

The solution to this is to pin futures.

Add a lint so that this won't be a problem going forward.

## Test plan

CI

## Migration plan

N/A

## Observability plan

N/A

## Documentation Changes

N/A