[DOC] Add overview to intro and tweak one sentence (#4967)
## Description of changes

This PR adds a bit of text to the intro and design sections to help the
reader who is coming in without any context to understand what wal3 can
do and how it's designed at a high level.

It also rephrases one sentence to avoid the word "comprise" because that
word is in the middle of linguistic drift, meaning both "the list of
parts of a whole" and it's opposite, "the whole that these parts compose
together." As a result this word is confusing to many, and rephrasing
without it generally improves both readability and clarity.

It includes no new code and no code change. Therefore, it also includes
no tests.

## Documentation Changes

The changes are only in the README, so no docstrings were changed.

The changes should correspond to the way the software already works (but
please double-check).