[CHORE]  Log a metric to track when the log needs to purge dirty. (#5482)
## Description of changes

In #5476 we performed read repair on collections that were in a bad
state.  That PR closes one hole, but leaves the same case open in a way
that requires operator intervention.  This PR adds a metric so we know
when the unaddressed condition happens.

## Test plan

CI

## Migration plan

N/A

## Observability plan

This.

## Documentation Changes

N/A