[BUG]  The total_records_post_compaction column could be inaccurate.
This column is only used for SPANN metrics, so the impact is limited for
now.  On next write the count from the segment will be used to
reinitialize such a collection.

Sequence of events triggering bug:
- Add records
- Compact everything
- Add records again, but nothing new.
- Compact everything

total_records_post_compaction would be zero because it is initialized to
zero in the orchestrator and the second materialize logs turns into a
NOP.