[BUG]  Distance calculations sporadically fail on sse.
The distance_sse functions started flaking in CI.  It looks like they
are correct, just not the same stability as the baseline.  Rather than
play a game of numerical stability to fix, I'm making the tolerance
proportional to vector length and thus the error.  Also, up the number
of iterations so that failures are >50% likely rather than 1 in 100.