[BUG]  The total_records_post_compaction column could be inaccurate. (#5497)
## Description of changes

This column is only used for SPANN metrics and stateful quota.
On next write the count from the segment will be used to
reinitialize such a collection.

Sequence of events triggering bug:
- Add records
- Compact everything
- Add records again, but nothing new.
- Compact everything

total_records_post_compaction would be zero because it is initialized to
zero in the orchestrator and the second materialize logs turns into a
NOP.

## Test plan

CI

## Migration plan

We need to decide if this column is necessary for SPANN.

## Observability plan

N/A

## Documentation Changes

N/A