[ENH]  Improve our usage of the AWS S3 SDK.
This PR makes it so that we introduce two new configurations:  retry
count and stall_protection_ms.  The stall protection should be 2/3 of
the retry_timeout_ms according to examples gleamed from the Internet.
Certainly less so that it can kick in before the retry.

The current retry_timeout_ms is converted into a per-attempt value to
maximally preserve current behavior.