[ENH]  Improve our usage of the AWS S3 SDK. (#5506)
## Description of changes

This PR makes it so that we introduce two new configurations:  retry
count and stall_protection_ms.  The stall protection should be 3/4 of
the retry timeout per attempt according to examples gleamed from the Internet.
Certainly less so that it can kick in before the retry.

The current retry_timeout_ms is converted into a per-attempt value to
maximally preserve current behavior.

## Test plan

CI + Staging

## Migration plan

N/A

## Observability plan

Observe and tune the values on staging.

## Documentation Changes

N/A