[ENH] Add support for not compacting metadata if indexing not enabled in schema (#5550)
## Description of changes

This PR adds piping of schema through compaction, via the
MetadataWriter. During metadata compaction, it checks against the schema
for whether a given key at a specific MetadataValueType is found in the
schema with indexing enabled. If so, it keeps that, and skips adding to
the metadata index if the it is not enabled. To reduce overhead of data
copying, a new enum MetadataValueType is introduced, derived form the
MetadataValue enum, but maps the value only to its type. This way,
during the pruning process of metadta keys to index, the metadata value
itself does not need to be transferred.
- Improvements & Bug fixes
  - ...
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility
changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_