[ENH]  Control how far into the future s3heap scans (#5620)
## Description of changes

We need to not fetch more than a minute or two into the future to make s3heap-service performant and
cost-effective.  This PR makes it so that the s3heap will allow limiting how far into the future a
call will look for peek/prune.

## Test plan

CI

## Migration plan

N/A

## Observability plan

N/A

## Documentation Changes

N/A