[ENH] Update fork log validation (#5778)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- Update the fork log validation logic. We only validate that the
compaction cursor is in the range of copied logs if it exists.
Previously, we enforce this validation by assuming the compaction cursor
is 1, which is not likely the case.
- New functionality
    - N/A

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility
changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the_ [_docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_