[BUG] is_default_schema does not do a space check for defaults or #embedding (#5787)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- There is a bug in schema where is_default_schema does not check the
space when confirming if the schema is default. This was hidden by our
e2e tests, since we used an ef with a default space of cosine. therefore
the configuration was cosine and was automatically converted into the
schema and tests passed. this PR adds a new test to confirm the basic
space setting scenario works
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility
changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_