[BUG] fix sparse autoembed queries on search api (#5800)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- There was a bug in the search api, where search queries were not
auto-embedded. This PR fixes that and adds a test to ensure that search
api explicitly works on autoembedding in both the read and write path.
previously, there were tests on the underlying search class and manually
calling the embed function to ensure the embedded results were valid.
this test is e2e, ensuring that adds and queries return the desired
result and the search api functions as intended
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility
changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_