[ENH]: Persist a backfill record on log to trigger backfill (#5897)
## Description of changes

_Summarize the changes made by this PR._

This diff persists a backfill signal that goes from a call to `attach_function`​ to `run_backfilled_attached_function_workflow`​, introduced by #5896 . This diff achieves that by pushing a new `backfill_fn`​ log to the concerned input collection's logs. This log should effectively be a no-op on the compactor path as materialize_logs makes sure to ignore these logs. In compact.rs::compact(), if the initial `run_fetch_logs`​ call detects the presence of this backfill log, `run_backfilled_attached_function_workflow` will be called. If this call succeeds, we return and ignore the usual compaction flow. If backfill was not needed, the usual pre-existing compaction + attached function execution is called.

- Improvements & Bug fixes
    - ...
- New functionality
    - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js, `cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need to make documentation changes in the_ [_docs section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_