[BUG]: Propagate flush failures to upstream with direct_hnsw (#5923)
## Description of changes

_Summarize the changes made by this PR._

- Improvements & Bug fixes
- There was a bug with direct_hnsw where if the underlying storage fails
to write the hnsw file then also it returns Ok() to upstream. This leads
to compactor falsely registering the unflushed file paths and subsequent
queries failing because the hnsw index does not exist on storage.
- Fixed by properly propagating the error and also added a unit test
that fails with the current code but passes with the fix
- New functionality
  - ...

## Test plan

_How are these changes tested?_
Added a unit-test
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Migration plan
None

## Observability plan
None

## Documentation Changes
None