[BUG]  Idempotent creates of attached functions fail.
Idempotent creates of attached functions fail because we check the error
and then throw an error anyway.  This PR adds a carve out for that error
condition and adds a test of the case that got us into the state where
this was noticed.

retrieve any (ready/not-ready) task

Rename ReadyOrNotReady