[ENH][wal3]  Pave the way for relative-offset parquet files.
In order to run replicated wal3, we need to be able to upload a fragment
without knowing its offset, but knowing that it occupies a contiguous
range of the log positions.  To do this, we allow the parquet file to
have an `offset` column or a `relative_offset` column.

When reading the parquet if there is an `offset` column, it gets used
immediately.

If there is no `offset` column, but an optional base is provided to the
reader and there is a `relative_offset` column, the relative offset will
be made absolute.

If neither case holds, we return an explicit error.