[BUG]: Fix deadlock in orchestrator test (#6020)
## Description of changes

`test_operator_cancellation` had a race condition before this change. It spawned a dispatcher with a max size of one with 2 worker threads before sending a task to it. If the task gets sent before the worker threads are able to send a `TaskRequestMessage` to this queue then the worker threads don't get registered with the dispatcher and there is nobody to process the enqueued task.

This diff fixes that by bumping the dispatcher size to 3. Also, this test does not seem to test cancellation anymore so it has been renamed.

- Improvements & Bug fixes
  - ...
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js, `cargo test` for rust

## Migration plan

_Are there any migrations, or any forwards/backwards compatibility changes needed in order to make sure this change deploys reliably?_

## Observability plan

_What is the plan to instrument and monitor this change?_

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need to make documentation changes in the [docs section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_