[DOC] Fix doc 500 due to tabs component (#5989)
We're getting 500s on several pages because the Tab markdoc component only has one child. This PR prevents the Tab component from breaking when it only has one child.

https://docs.trychroma.com/integrations/embedding-models/text2vec
https://docs.trychroma.com/integrations/embedding-models/nomic
https://docs.trychroma.com/integrations/embedding-models/open-clip


Additionally, this changes the persistent language behavior for tabs. Before, if you chose "Typescript" and then went on this page where there is no Typescript option, it would simply have no tab selected and show nothing. 

Now, if the preferred tab is not available, it will default to the first tab option. This is at the cost of the tabs not updating to your new preferred language while you're on the page -- only when you change pages. 

![Screenshot 2025-12-08 at 3.21.09 PM.png](https://app.graphite.com/user-attachments/assets/aabe734c-d783-47ab-a081-44a14dd40f71.png)