[BUG] BM25 does not work when multithreading (#5993)
closes #5969

`ChromaBM25EmbeddingFunction` uses `snowballstemmer` to tokenize.
Currently, all invocations of `ChromaBM25EmbeddingFunction` share the
same `snowballstemmer` instance because it's created in `__init__` and
wrapped in a `@cache`.

The issue is that `snowballstemmer` instances are not thread-safe. The
tokenizer has an internal state. If you use the same `snowballstemmer`
object to tokenize concurrently, it will break.

```python
class EnglishStemmer(BaseStemmer):
    '''
    This class implements the stemming algorithm defined by a snowball script.
    Generated from english.sbl by Snowball 3.0.1 - https://snowballstem.org/
    '''

    g_aeo = {u"a", u"e", u"o"}

    g_v = {u"a", u"e", u"i", u"o", u"u", u"y"}

    g_v_WXY = {u"a", u"e", u"i", u"o", u"u", u"y", u"w", u"x", u"Y"}

    g_valid_LI = {u"c", u"d", u"e", u"g", u"h", u"k", u"m", u"n", u"r", u"t"}

    B_Y_found = False
    I_p2 = 0
    I_p1 = 0

    def __r_prelude(self):
        self.B_Y_found = False # <-- mutates state
 ...
```

Claude wrote me a small benchmark to measure how impactful it is to
instantiate a new `snowballstemmer`.

```
Benchmarking Stemmer Creation Time

1. Direct _SnowballStemmerAdapter creation:
   Mean: 0.0004 ms
   Median: 0.0003 ms
   Min: 0.0002 ms
   Max: 0.0096 ms
   Std Dev: 0.0009 ms
```

It's insignificant enough that I think it's reasonable to instantiate a
new object per-call.

## Testing

I created a new test for BM25EF. It does not pass on `main`, but it
passes with my changes.