[DOC] Enhance JS reference docs (#6062)
I added the new Search API to the docs

Additionally, I expanded the params to be useful. For example, this is
how the docs currently look like.

![Screenshot 2025-12-17 at
4.46.36 PM.png](https://app.graphite.com/user-attachments/assets/e4a65bc9-ff33-4c42-94e2-553b4d37ffa7.png)

This doesn't really help me because what I really want to know is what
the params are. I expanded the tables to this:

![Screenshot 2025-12-17 at
4.46.55 PM.png](https://app.graphite.com/user-attachments/assets/81b30280-822c-4077-8252-a3e192f161f8.png)